Imported MQTT components
This commit introduces a 'ext' directory which incubates components that
reasonably could be promoted to their own projects.

Formerly Jig's MQTT components have been imported and migrated to be
compatible with Stuart Sierra's component library.

A new protocol (NettyHandlerProvider) has been introduced for Netty
handler component to satisfy. This is a better integration point than
the convention of requiring components create a ::handler-factory entry
in the system map.