Various fixes towards a working system
Taken out remaining references in mqtt.clj to the global system (there
is no global system anymore, only components and their state).

Added a priority function to the NettyProviderHandler because the
handlers must appear in the Netty chain of handlers in the correct order
and the component library is non-deterministic with respect to
dependency ordering. The priority function is used to sort the handlers first.